TextInputWayland: Fix 2 text input bugs
Fixes two bugs related to text input processing. The first is that the
input could get stuck in an infinite loop of canceling the composition
if the Editor though there was a composition, but the text input handler
didn't. The second is that some compositors send a leave event when the
text input is deactivated, but this was getting processed after a new
text input was activated (i.e. when focus shifted directly from one text
field to another). This delay caused the new text input to be
immediately deactivated, which was undesirable.