TextInputWayland: Always commit on commit event
Fixes a bug where the commit string was not committed when a commit
event was received from the Wayland compositor. Typically, this
manifested as nothing happening if a commit event was sent without any
previous preedit string. The reason this happened is that the internal
Editor state would only report that a commit was possible if a preedit
string was present. This is incorrect, as we always want to commit as
long as we have a non-empty commit string.