Build failure when ENABLE_VIDEO ENABLE WEB_AUDIO are disabled
USE_GSTREAMER is defined as PRIVATE in the cmake. The default value is
set to ON in `Source/cmake/GStreamerDefinitions.cmake`.

This settings was introduced in the commit e1152126e4af

The USE_GSTREAMER is never evaluated during configuration time and it
is assigned and set as True (ON) only in the
`Source/cmake/GStreamerChecks.cmake`
file when ENABLE_VIDEO or ENABLE_WEB_AUDIO are also True. As the
default value and the asignation set the very same value, the default
value of this variable will be always True.

It seems USE_GSTREAMER is only interesting when video or audio stuffs
are activated, then I'd suggest to set the default value as OFF and
relay in the activation of this varaible done in
`Source/cmake/GStreamerChecks.cmake`. Note that this is the case in WPE
upstream: USE_GSTREAMER is not set to ON in
`Source/cmake/GStreamerDefinitions.cmake`.

Ref:
https://github.com/WebPlatformForEmbedded/WPEWebKit/commit/e1152126e4af