Decrease DisplayRefreshMonitorTimer priority
DisplayRefreshMonitorTimer priority is used for timers that schedule callbacks in
requestAnimationFrame api. They have higher priority than regular timers
scheduled in setTimeout api. On pages with lots of animations the result
is that setTimeout timers are starved out and can execute seconds after
they're supposed to (sometimes even tens of seconds).

This commits makes sure both priorities are the same.