[ARRISEOS-41835] UIProcess should not block on IPC
UIProcess should process incoming sync IPC from WebProcess when waiting for a
sync IPC reply from it.
Result of this bug is a timeout error coming from waitForSyncReply()
function.
This patch is based on below upstream fix:
https://bugs.webkit.org/show_bug.cgi?id=189927
https://github.com/WebPlatformForEmbedded/WPEWebKit/commit/005814e6033e7d9390e42aff26cc1abc291c00a5

Unfortunately that fix is based on refactoring in WebPageProxy.cpp:
    Refactoring related to Safe Browsing
    https://bugs.webkit.org/show_bug.cgi?id=189631
and doesn;t apply cleanly, so it was modified.