Don't evict the same samples range multiple times
Improve moving coded frames removal window to skip
already removed parts and start from the end of previous
removal.
The only 'real' change here is termination condition
for loop that removes data ahead of current time
(some time ranges was not removed if rangeStart < minRangeStart
and rangeEnd > minRangeEnd).
All the rest improves looping but the internal flow
is exactly the same (the same set of samples would be removed)

E.g. Having start of buffered range at 100sec and maximumRangeEnd and 200sec:
was:
1) (0-100)
2) (30-130)
3) (60-160)
...
will be:
1) (0-100)
2) (100-130)
3) (130-160)
...

The same for removing sampleas ahead of currentTime (with termination point
fixed)