Round SourceBuffer removal range
Round SB removal range provided as 'double' in SB::remove() func
to 1/1000 us.

This helps to avoid cases when unintended frame
would be marked for removal becasue of very small overlap
that could come from precision lost.

It fixes HBOMax issue that uses next sample PTS as removal 'end'
and unexpectedly removes the next segment (as it's a sync frame)