Limit queue high-watermark for REALTEK platform only
This narrows https://github.com/WebPlatformForEmbedded/WPEWebKit/pull/953
to REALTEK platform only as the source of original issue.

queue2 by default uses max 2sec of buffering (estimated from
input data bitrate) and setting high-watermark as 0.1 limits this
to 0.2sec that shinks the buffer to be very small.

As a result queue2 is constantly sending buffering messages
(especially at the begining of playback) with random values 0-100%
as it's very easy for the pipeline to completely drain the queue.

REALTEK platform decoder doesn't seem to limit audio buffering enough
so for audio live streams it is taking a lot of time to fill up.