Cache video quality stats
When the pipeline is composed by a playsink, setting the
pipeline to the READY state (e.g. when reaching EOS) may
cause the videosink state to be set to NULL as part of
the videochain deactivation process (videosink is
a child element of playsink). Some sink implementations
reset the stats when the element's state is set to NULL,
but the application may still query them as from app's
perspective they should be still retrievable.
Keep a cache of latest stats and return cached values
when sink state is NULL