[GStreamer] Instantiate downloadbuffer element with playbin3 if video preload="auto" https://bugs.webkit.org/show_bug.cgi?id=261737
Reviewed by Philippe Normand.

When using WEBKIT_GST_USE_PLAYBIN3=1 and cog to open website with content
<video src="./test.mp4" preload="auto" autoplay="autoplay" loop></video>
the video would not be buffered into on-disk cache as it used to be with
playbin(2). This is because playbin3 uridecodebin instantiated urisourcebin
does not instantiate downloadbuffer element by default, this has to be
explicitly enabled by setting the urisourcebin element 'download' parameter
to TRUE.

Set the urisourcebin element 'download' parameter to TRUE in case of playbin3
and non-MSE non-MediaStream sources which use '<video ... preload="auto">' .
Note that the 'preload="auto"' is the default as well.

* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::handleStreamCollectionMessage):

Canonical link: https://commits.webkit.org/268335@main