[GStreamer] Ignore notify::caps with NULL caps https://bugs.webkit.org/show_bug.cgi?id=262949
Reviewed by Philippe Normand.

notify::caps is triggered not only when caps are set, but also when
they're unset, as it's the case during pipeline teardown.

This was happening as the state pipeline was set to NULL from the main
thread, making the non-main-thread assertion
AbortableTaskQueue::enqueueTask() fail.

* Source/WebCore/platform/graphics/gstreamer/TrackPrivateBaseGStreamer.cpp:
(WebCore::TrackPrivateBaseGStreamer::installUpdateConfigurationHandlers):

Canonical link: https://commits.webkit.org/269307@main