Fix return of invalid playback position
After a seek, the gstreamer pipeline may not be able to return a
valid position right away.
The m_canFallBackToLastFinishedSeekPosition is reset in an async
way when handleMessage() is called, so playbackPosition() might
be called while we still not have a valid position from the
pipeline.

This also takes care of a corner case where the cached position
invalidation task (async operation triggered by the call to
invalidateCachedPositionOnNextIteration()) runs after a seek
operation completing, but without a valid position being
yet available from the pipeline. In such case, the cached
position is set with seeked-to position, but since the
position invalidation request has already been requested,
 it would now invalidate an actual valid position.