[VisionOS] Twitter.com videos do not play https://bugs.webkit.org/show_bug.cgi?id=267938 rdar://121391975
Reviewed by Tim Horton.

Twitter.com has adopted ManagedMediaSource, and disables airplay via
HTMLMediaElement.disableRemotePlayback. On other platforms, this allows
MMS to move to the "open" state, due to the managedMediaSourceNeedsAirPlay
setting. However, on VisionOS, the RemotePlaybackEnabled setting is off,
so the disableRemotePlayback property is not reflected into an attribute
on the HTMLMediaElement, and the check for isWirelessPlaybackTargetDisabled()
fails.

Selectively move portions of EnabledBySetting=RemotePlaybackEnabled from
the HTMLMediaElement+RemotePlayback.idl file sufficient for disableRemotePlayback
to be reflected, and to allow ManagedMediaSource to check for its presence.

* Source/WebCore/Modules/remoteplayback/HTMLMediaElement+RemotePlayback.idl:

Canonical link: https://commits.webkit.org/273434@main