Issue: Audio not heard after seek of AppleTV video
Description: When AppleTV+ video content is played and if seek/jump is
performed, it is seen that audio is not heard. AppleTV uses AC3
encrypted audio. AC3 parser which is automatically added in the gst
pipeline drops encrypted audio.

Solution: Make AC3 parser lowest rank so as to not get included in the
gst pipeline.