[2.38] limit conservative scan range
Turns out that conservative GC root scanning is going too deep into the stack
and this may prevent the collection of objects - see:

https://github.com/WebPlatformForEmbedded/WPEWebKit/issues/1363

The change introduces 'stack guards' that prevent this in particular cases,
that is - if there is gloop main loop routine in the stack the GC root stack search
will not go deeper that this frame. This avoids some 'false positives' conservative roots
that can make the app hold a lot of memory.