[MSE][GStreamer] Support markEndOfStream() before appendBuffer()
MediaSource::markEndOfStream() causes the SourceBuffer TrackBuffers to push
EOS to playback pipeline through WebKitMediaSrc. When that happens without
any SourceBuffer::appendBuffer() call, the pipeline can't finish autoplugging
and parsebin triggers an unrecoverable error.

While it's certainly possible to change parsebin to not trigger the error,
the truth is that this seems legitimate GStreamer behaviour that shouldn't be
altered (and also, doesn't fix the problem by itself, I've checked it).

This commit adds support for temporarily ignoring the error and asking the
pipeline to change to READY state. It also notifies HTMLMediaElement about
timeChanged, so the "ended" event can be triggered (after all, there's no
other position to go beyond 0, as there are no samples and 0 is already the
duration, so technically playback is ended).