[HTMLMediaElement] Reschedule `timeupdate` event if fired too early
HTMLMediaElement fires `timeupdate` every 250ms in repeating timer.
If timer is fired before 250ms from the last occurance it is silently skipped
and needs to wait for another 250ms. As a result the gap between
two following timeupdate events may vary 250-500ms.
This may happen in two cases:
1) Non-periodic timeupdate event is scheduled for any reason
2) When the difference between two timer calls is lower than 250ms.
The second may happen when there are multiple timers to handle at the same time
in ThreadTimers (maxDurationOfFiringTimers)