Add GStreamer quirk for loading extra system plugins with gstreamer-full
This patch adds an extraSystemPlugins() method to the GStreamer quirks,
which is responsible for specifying which plugins to be loaded from the
system GStreamer installation, pointed to by
GST_PLUGIN_SYSTEM_PATH{_1_0}.

This allows us to use a recent gstreamer version with gstreamer-full containing
only the necessary libraries and plugins that aren't hardware dependant, while
still allowing specific hardware-accelerated plugins to be loaded from
the system gstreamer installation.

That way, we aren't required to rebuild hardware-accelerated plugins
with gstreamer-full, and the relevant plugins that are present in system with
quirks can still be loaded even if they were built for an older
GStreamer version, since the ABI for the plugins is stable.

In this patch, the new method is implemented for broadcom and westeros,
and add a new quirk is added for raspberrypi to load the OMX plugin
(although this new quirk should probably be removed and we should move
to V4L2 instead).