Use SSLKEYLOGFILE env to enable logging SSL keys in libwebrtc / boringssl
For debugging, it's sometimes useful to be able to decrypt captured TLS traffic.
By setting SSLKEYLOGFILE env variable, it's possible to log secrets to a file,
which can then be used for decryption.
But libwebrtc uses a separate TLS implementation, boringssl, which does not support it by default.

Add key logging in libwebrtc if the env is set.