StringImpl refcounting should be atomic https://bugs.webkit.org/show_bug.cgi?id=289962 rdar://147313174
Backport notes:
I removed some optimizations; ref churn is not great, but still correct.

Original notes:

Reviewed by Antti Koivisto, Yusuke Suzuki, Chris Dumez, and Keith Miller.

This patch takes the first step to making WTF::String thread-safe: Making
m_refcount increment and decrement atomic. (Hash, flags, and the AtomString
table are not yet atomic. I plan to address those in my next patch.)

The direct code change is pretty small, but it measured as a performance
regression. So most of this code is optimization. I got rid of about 1/3 of
WTF::StringImpl refcounts on Speedometer and about 1/2 on JetStream.

ARM has a special ldadd instruction for fast atomics. Net net, it measures as
about 1.5X slower than just add. That's pretty good! But it still required
some optimizations to be affordable. Intel atomics are much slower, so it took
far more optimizations to get Intel across the finish line.

A/B testers suggest that we might still have a < 0.5% JetStream regression on
Intel. I checked in with Yusuke, Keith, and Maciej, and we can move forward
because we have a large total JetStream speedup on Intel this year, so customers
on older hardware will still see a net speedup when they get new software.