[ARMv7] Preserve return address without ISA mode bit https://bugs.webkit.org/show_bug.cgi?id=299515
Reviewed by Yusuke Suzuki.

On ARMv7, the least significant bit in the link register indicates the
mode (ARM/Thumb) of the code that we return to. Callers of
preserveReturnAddressAfterCall care about the return address and when
calling the throwExceptionFromOMGThunk (as introduced in 300298@main)
the LSB being set results in a return address outside the function,
whose CallSiteIndex can therefore not be looked up.

Simply clear the LSB when preserving the return address on ARMv7.

Canonical link: https://commits.webkit.org/300555@main

(cherry picked from commit 4d672bf7ca92bcdf9ceac15576d4443c36343476)